Split CSV download into separate module
